Replace AS merger with AnimSeq merger
