+++ crates/uk-content/src/actor/params/as/res_skel_asset.rs
