+++ crates/uk-content/src/actor/params/mod.rs
@@ -22,4 +22,3 @@ pub mod rgconfig;
 pub mod rgconfiglist;
 pub mod shop;
 pub mod umii;
-pub mod anim_seq;
